server/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
bcrypt==4.1.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
ecdsa==0.18.0
fastapi==0.110.0
filelock==3.13.1
flatbuffers==24.3.7
fsspec==2024.3.1
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.6
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
passlib==1.7.4
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyPDF2==3.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.2
SQLAlchemy==2.0.29
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
tf_keras==2.16.0
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
